Make substTt sound
substTt, which substitutes a subexpression for a variable in a TL
expression, did not perform capture-avoiding substitution and was thus
unsound.

Example. Substituting 'const 4.2' for 'a' in 'λx. x + a' works fine:

> substTt "a" (Op (Constant 4.2) Unit) TScal (Lambda "x" TScal (Op EScalAdd (Pair (Var "x" TScal) (Var "a" TScal))))
\x -> (evalOp EScalAdd (x, evalOp const(4.2) ()))

But substituting 'x' for 'a' in 'λx. x + a' incorrectly gave 'λx. x + x':

> substTt "a" (Var "x" TScal) TScal (Lambda "x" TScal (Op EScalAdd (Pair (Var "x" TScal) (Var "a" TScal))))
\x -> (evalOp EScalAdd (x, x))

whereas the correct answer would have been 'λy. y + x' for some new,
fresh variable 'y'. The new code performs capture-avoiding substitution.