Pointful linear representation
@VMatthijs, what do you think about this? Look at the top of
TargetLanguage.hs, where linear functions in the target language are
represented in a _pointful_ format that corresponds much more closely to
the paper.

Downside of this currently is that the reverse AD result of the first
example in the paper (which, of course, is an example for forward AD,
but that aside) becomes this after simplification:

let x1 = 2.0
    x2 = x1 * ctxtVar1
    x3 = ctxtVar1 * x2
 in ( ((x2, x3), cos x3)
    , \v ->
        let v =
              plus
                (zero, fst (fst v))
                (let v = plus (snd (fst v)) (lscalprod (0.0 - sin x3) (snd v))
                  in ((zero, lscalprod x2 v), lscalprod ctxtVar1 v))
         in plus (fst v) (zero, lscalprod x1 (snd v)))

which is not bad, but that nested let-binding in the linear function is
awkward. Even more so because my current pointful linear language only
allows "shadowing let" using LinLet. I'll see if I can make it work
using a linear environment instead of just a linear variable.