Forward AD
@VMatthijs This uses some wild environment tricks to un-linearise
TargetLanguage.LinTTerm.LinLam in ToConcrete.toConcreteL' . What happens
here is the following:

- I'm converting a term of type 'LinTTerm env lenv (a -> t)', which is
  of the form 'LinLam t' for 't :: LinTTerm (a : env) lenv t'.

- I'm converting it _to_ a concrete term of type:
    'CTerm (UnLinEnv (lenv ++ env)) (UnLin a -> UnLin t)'
  which is going to be a CLambda with a body of type:
    'CTerm (UnLinEnv (a : lenv ++ env)) (UnLin t)'.
  (Note that 'UnLinEnv' is just 'map UnLin', except no dependent types.)
  (Note that the variables bound _last_, so innermost, are at the _head_
   of the environment list in my case. The point of this 'lenv ++ env'
   type is that the environment of the term being produced has its
   unrestricted variables first, on top, and then the linear variables
   at the bottom of the environment, where stuff gets pushed onto it.)

- The whole point of this LinLam constructor is to create an
  _unrestricted_ lambda, and we see this in the types: we're pushing an
  element on the unrestricted variables stack ('env').

- But the concretised version of our body expects this new unrestricted
  variable (of type 'a') to be at the bottom of the _unrestricted_ part
  of the stack! This means that we'll have to move this variable past
  all the linear variables in scope.

And this "moving past" is exactly what the wOverEnvUnLin weakening
action does. Applying that action to a term turns it into a term that
expects the new unrestricted variable at the bottom of the entire stack
instead of halfway down. It needs to explicitly use UnLin right there in
the type class because I can't "just" convert an 'env1 :> env2' to an
'UnLin env1 :> UnLin env2'. (I think this has to do with UnLin not being
an injective type family, which prevents me from writing the obvious
spelled-out identity function. Not sure exactly.)

The separate 'go' function is literally only so that I get the 'a' type
variable in scope (with ScopedTypeVariables) for the '@a' visible type
application.

It works; the two forward AD examples in the paper give the expected
result. But I'd like this to be easier than it is.